adding gui start
