+++ src/lib.rs
@@ -1,4 +1,8 @@
 //! Welcome to the code-documentation of flakeshot!
+
+use frontend::main_window::AppModel;
+use gtk4::CssProvider;
+use relm4::RelmApp;
 pub mod backend;
 pub mod cli;
 pub mod frontend;
@@ -17,4 +21,23 @@ pub mod tray;
 pub enum Error {
     #[error("An error occured in the backend: {0}")]
     Backend(#[from] backend::Error),
+}
+
+pub fn start_gui() {
+    let app = RelmApp::new("org.flakeshot.app");
+    relm4_icons::initialize_icons();
+    initialize_css();
+
+    app.run::<AppModel>(());
+}
+
+fn initialize_css() {
+    let provider = CssProvider::new();
+    provider.load_from_data(include_str!("frontend/style.css"));
+
+    gtk4::style_context_add_provider_for_display(
+        &gdk4::Display::default().unwrap(),
+        &provider,
+        gtk4::STYLE_PROVIDER_PRIORITY_APPLICATION,
+    );
 }
